+++ assets/src/scripts/gutenberg.js
@@ -1,6 +1,4 @@
-/**
- * Import Gutenberg custom blocks.
- */
+// Import Gutenberg custom blocks.
 require( './blocks/small-cta' );
 require( './blocks/large-cta' );
 require( './blocks/cta-template' );
@@ -9,3 +7,6 @@ require( './blocks/latest-lessons' );
 require( './blocks/youtube-post-template' );
 require( './blocks/ypt-search' );
 require( './blocks/section' );
+
+// Import utils to customize Gutenberg interface.
+require( './modules/gutenbergUtils' );